+++ packages/expo-cli/commands/upload-sourcemaps.js
@@ -2,7 +2,7 @@ const prompts = require('prompts')
 const addPlugin = require('../lib/configure-plugin')
 const installPlugin = require('../lib/install-plugin')
 const { onCancel, getDependencies } = require('../lib/utils')
-const { isEarlierVersionThan } = require('../lib/version-information')
+const { isEarlierVersionThan, getBugsnagVersionForExpoVersion } = require('../lib/version-information')
 const { blue, yellow } = require('kleur')
 
 const PLUGIN_NAME = '@bugsnag/plugin-expo-eas-sourcemaps'
@@ -23,7 +23,7 @@ module.exports = async (argv, globalOpts) => {
   const res = await prompts({
     type: 'confirm',
     name: 'addPlugin',
-    message: 'Do you want to automatically upload source maps to Bugsnag? (this will modify your app.json)',
+    message: 'Do you want to automatically upload source maps to Bugsnag? (this will modify your app.json and package.json)',
     initial: true
   }, { onCancel })
 
@@ -38,7 +38,12 @@ module.exports = async (argv, globalOpts) => {
         yarn: globalOpts.yarn
       }
 
-      await installPlugin(projectRoot, options)
+      // install a plugin version that matches the SDK version, i.e. SDK 48 -> @bugsnag/plugin-expo-eas-sourcemaps@^48.0.0
+      // if there is no suitable bugsnag version we haven't yet released support for this Expo version, so install the latest
+      const versionInformation = getBugsnagVersionForExpoVersion(installedExpoVersion)
+      const pluginVersion = versionInformation ? versionInformation.bugsnagVersion : 'latest'
+
+      await installPlugin(pluginVersion, projectRoot, options)
     }
 
     console.log(blue('> Inserting EAS plugin into app.json'))