+++ packages/plugin-expo-eas-sourcemaps/src/android.js
@@ -1,28 +1,13 @@
-const {
-  AndroidConfig,
-  withAppBuildGradle,
-  withAndroidManifest
-} = require('@expo/config-plugins')
+const { AndroidConfig, withAndroidManifest } = require('@expo/config-plugins')
 
 // Using helpers keeps error messages unified and helps cut down on XML format changes.
 const { addMetaDataItemToMainApplication, getMainApplicationOrThrow } = AndroidConfig.Manifest
 
 const withAndroidPlugin = (config, onPremConfig) => {
-  // Update android manifest with bugsnag config
+// Set the Bugsnag API key in the android manifest - note that this isn't required for source map uploads
+// but is added here to support reporting builds using the fastlane plugin
   config = withAndroidManifest(config, config => {
     config.modResults = setBugsnagConfig(config, config.modResults)
-    return config
-  })
-
-  // Inject gradle dependencies
-  config = withAppBuildGradle(config, (config) => {
-    if (config.modResults.language === 'groovy') {
-      config.modResults.contents = injectDependencies(config.modResults.contents)
-    } else {
-      throw new Error(
-        'Cannot configure Bugsnag in the app gradle because the build.gradle is not groovy'
-      )
-    }
     return config
   })
 
@@ -37,32 +22,15 @@ function setBugsnagConfig (config, androidManifest) {
   // Get the <application /> tag and assert if it doesn't exist.
   const mainApplication = getMainApplicationOrThrow(androidManifest)
 
-  addMetaDataItemToMainApplication(
-    mainApplication,
-    apiKeyName,
-    apiKeyValue
-  )
+  if (apiKeyValue) {
+    addMetaDataItemToMainApplication(
+      mainApplication,
+      apiKeyName,
+      apiKeyValue
+    )
+  }
 
   return androidManifest
-}
-
-function injectDependencies (script) {
-  return `buildscript {
-        repositories {
-            mavenCentral()
-        }
-        dependencies {
-            classpath 'com.bugsnag:bugsnag-android-gradle-plugin:[7.3.0,)'
-        }
-    }
-
-    ${script}
-
-    apply plugin: 'com.bugsnag.android.gradle'
-
-    bugsnag {
-        uploadReactNativeMappings = true
-    }`
 }
 
 module.exports = {